Added newline handlign

+++ src/drivers/screen.h
@@ -5,8 +5,8 @@
 #include "../types.h"
 
 #define VIDEO_ADDRESS 0xb8000
-#define MAX_ROWS 25
-#define MAX_COLS 80
+#define MAX_ROWS (u8)25
+#define MAX_COLS (u8)80
 
 #define REG_SCREEN_CTRL 0x3D4
 #define REG_SCREEN_DATA 0x3D5
